P5: ERBench novel families (9,118 entries, 5 catalogs) + SA firewall + sd 0.12.4
- erbench-syneq@1 (5301, MIT) / erbench-phybench@1 (90, MIT) / erbench-densities@1 (33, BSD-3)
  published to the main HF repo; erbench-oeis@1 (3499) + erbench-eponymous@1 (195) published to
  the SEPARATE CC-BY-SA repo psaegert/symbolic-data-assets-sa (license card, LICENSE with full
  legal code + derivation statement, per-entry A-number/Wikipedia attribution; vendored SA csv
  drops the description column — upstream rows embed whole OEIS entry texts). Resolver's
  repo:name@version syntax verified for the SA repo.
- Canonicalization: '^'->'**', Abs->abs, paren-balanced two-arg log rewrite (E/10/other bases),
  pi/E -> numeric literals (house convention per feynman.yaml). 9,118/9,118 realize; audit:
  near-full validity everywhere (ERBench supports avoid invalid regions by design).
- sd 0.12.4: per-point rejection judges validity in the STORAGE dtype (float32) — two OEIS
  fast-growing formulas shipped inf y-values pre-fix (+ regression test).
- Re-rendering families dedup (Feynman/SRDS/Nguyen/Keijzer/Korns/Livermore/Vlad/Koza/Pagie);
  Strogatz skipped (ODE + GPL-3.0). Registry: 10,480 entries / 7,854 unique / 490 clusters.
- Follow-ups recorded: AIFeynman bonus set (import from original source), 3 extra Korns.

+++ src/symbolic_data/catalog.py
@@ -389,7 +389,11 @@ class ProblemCatalog(Catalog):
         # This only removes the exponential cost (f^-n vs 1/f draws) that made partial-domain
         # entries (e.g. sqrt(x1+x2) over a box straddling the domain) exhaust max_trials.
         # An attempts cap keeps truly degenerate entries (f ~ 0) on the honest placeholder path.
-        finite_mask = np.isfinite(x_all).all(axis=1) & np.isfinite(y_all).all(axis=1)
+        # Validity is judged in the STORAGE dtype (float32): a point whose y is finite in float64
+        # but overflows float32 (e.g. fast-growing integer-sequence formulas near their support
+        # edge) would otherwise ship as inf in the frozen Problem arrays.
+        finite_mask = (np.isfinite(x_all.astype(np.float32)).all(axis=1)
+                       & np.isfinite(y_all.astype(np.float32)).all(axis=1))
         if finite_mask.all() and n_first > n_points:
             x_all, y_all = x_all[:n_points], y_all[:n_points]
         elif not finite_mask.all():
@@ -413,7 +417,8 @@ class ProblemCatalog(Catalog):
                     yb = broadcast_target(evaluate(compiled, vb), batch, entry.id).reshape(-1, 1)
                 except Exception as exc:
                     raise NoValidSampleFoundError(f"evaluation failed for {entry.id!r}: {exc}") from exc
-                mask = np.isfinite(xb).all(axis=1) & np.isfinite(yb).all(axis=1)
+                mask = (np.isfinite(xb.astype(np.float32)).all(axis=1)
+                        & np.isfinite(yb.astype(np.float32)).all(axis=1))
                 keep_x.append(xb[mask])
                 keep_y.append(yb[mask])
                 collected += int(mask.sum())
